WIP explanation tests and retriving explanation impl
