+++ dependencies/graknlabs/dependencies.bzl
@@ -29,12 +29,12 @@ def graknlabs_grakn_core():
     git_repository(
         name = "graknlabs_grakn_core",
         remote = "https://github.com/graknlabs/grakn",
-        commit = "8d121f92b8ab2708399c6d57c30e074bb148ef3e" # sync-marker: do not remove this comment, this is used for sync-dependencies by @graknlabs_grakn_core
+        commit = "f80aee55f64d37073ef5ee2418c076ee1f23fd57" # sync-marker: do not remove this comment, this is used for sync-dependencies by @graknlabs_grakn_core
     )
 
 def graknlabs_protocol():
     git_repository(
         name = "graknlabs_protocol",
         remote = "https://github.com/graknlabs/protocol",
-        commit = "af2daa4b0ddce19b4668919ecccd059eb282292b" # sync-marker: do not remove this comment, this is used for sync-dependencies by @graknlabs_protocol
+        commit = "143b06e0559c8d52323e694fafd3a1b4169c969b" # sync-marker: do not remove this comment, this is used for sync-dependencies by @graknlabs_protocol
     )
